refactor(consensus): move EIP-8130 coinbase tip decode off the tx type
Keep TxEip8130 as the wire body and recover phase-0 tips through CoinbaseTip::decode so ordering analysis is not mixed into the protocol type.
