+++ crates/common/consensus/src/lib.rs
@@ -29,11 +29,11 @@ mod transaction;
 pub use transaction::serde_deposit_tx_rpc;
 pub use transaction::{
     AccountChange, AccountChangeChannel, BasePooledTransaction, BaseTransaction,
-    BaseTransactionInfo, BaseTxEnvelope, BaseTypedTransaction, Call, ChangeType, CreateEntry,
-    DEPOSIT_TX_TYPE_ID, Delegation, DepositInfo, DepositTransaction, EIP8130_REJECTION_MSG,
-    EIP8130_TX_TYPE_ID, Eip8130Constants, Eip8130Contracts, Eip8130Signed, Eip8130StaticError,
-    Eip8130TimestampError, IDefaultAccount, InitialActor, OpTxType, Scope, SignedAccountChanges,
-    SignedChange, TxDeposit, TxEip8130,
+    BaseTransactionInfo, BaseTxEnvelope, BaseTypedTransaction, Call, ChangeType, CoinbaseTip,
+    CreateEntry, DEPOSIT_TX_TYPE_ID, Delegation, DepositInfo, DepositTransaction,
+    EIP8130_REJECTION_MSG, EIP8130_TX_TYPE_ID, Eip8130Constants, Eip8130Contracts, Eip8130Signed,
+    Eip8130StaticError, Eip8130TimestampError, IDefaultAccount, InitialActor, OpTxType, Scope,
+    SignedAccountChanges, SignedChange, TxDeposit, TxEip8130,
 };
 
 mod extra;
